msm: mdss: use u64 instead of u32 for bitclock rate calculation
Bitclock calculation for high resolution panel overflows the
32bit variable which causes invalid pixel clock and byte clock
setting. Updating it to 64bit fixes the overflow issue.

+++ drivers/video/msm/mdss/mdss_dsi.h
@@ -204,7 +204,7 @@ extern u32 dsi_irq;
 extern struct mdss_dsi_ctrl_pdata *ctrl_list[];
 
 struct dsiphy_pll_divider_config {
-	u32 clk_rate;
+	u64 clk_rate;
 	u32 fb_divider;
 	u32 ref_divider_ratio;
 	u32 bit_clk_divider;	/* oCLK1 */
